refactored some more classes into main grakn-test artifact
