+++ grakn-test/src/test/java/ai/grakn/test/graql/graql/DeleteAndInsertTest.java
@@ -23,6 +23,7 @@ import ai.grakn.graphs.MovieGraph;
 import ai.grakn.graql.Var;
 import ai.grakn.graql.VarPattern;
 import ai.grakn.test.GraphContext;
+
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
