+++ grakn-test/src/test/java/ai/grakn/test/graql/parser/QueryToStringTest.java
@@ -26,6 +26,7 @@ import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.Query;
 import ai.grakn.graql.QueryBuilder;
 import ai.grakn.test.GraphContext;
+
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
