+++ grakn-test/src/test/java/ai/grakn/test/graql/printer/GraqlPrinterTest.java
@@ -28,6 +28,7 @@ import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.internal.printer.Printers;
 import ai.grakn.graql.internal.query.QueryAnswer;
 import ai.grakn.test.GraphContext;
+
 import org.apache.commons.lang.StringUtils;
 import org.junit.ClassRule;
 import org.junit.Test;
