+++ grakn-test/src/test/java/ai/grakn/test/graql/printer/JsonPrinterTest.java
@@ -26,6 +26,7 @@ import ai.grakn.graphs.MovieGraph;
 import ai.grakn.graql.Printer;
 import ai.grakn.graql.internal.printer.Printers;
 import ai.grakn.test.GraphContext;
+
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
