+++ grakn-test/src/test/java/ai/grakn/test/graql/query/AskQueryTest.java
@@ -21,6 +21,7 @@ package ai.grakn.test.graql.query;
 import ai.grakn.graphs.MovieGraph;
 import ai.grakn.graql.QueryBuilder;
 import ai.grakn.test.GraphContext;
+
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
