+++ grakn-test/src/test/java/ai/grakn/test/graql/query/QueryBuilderTest.java
@@ -25,6 +25,7 @@ import ai.grakn.graql.DeleteQuery;
 import ai.grakn.graql.InsertQuery;
 import ai.grakn.graql.MatchQuery;
 import ai.grakn.test.GraphContext;
+
 import org.junit.After;
 import org.junit.ClassRule;
 import org.junit.Rule;
