+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/inference/GenealogyTest.java
@@ -27,6 +27,7 @@ import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.admin.MatchQueryAdmin;
 import ai.grakn.graql.internal.reasoner.query.QueryAnswers;
 import ai.grakn.test.GraphContext;
+
 import com.google.common.collect.Sets;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
