+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/inference/GeoInferenceTest.java
@@ -27,6 +27,7 @@ import ai.grakn.graql.QueryBuilder;
 import ai.grakn.graql.Var;
 import ai.grakn.graql.internal.reasoner.query.QueryAnswers;
 import ai.grakn.test.GraphContext;
+
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
