+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/inference/RecursiveInferenceTest.java
@@ -35,6 +35,7 @@ import ai.grakn.graql.MatchQuery;
 import ai.grakn.graql.QueryBuilder;
 import ai.grakn.graql.internal.reasoner.query.QueryAnswers;
 import ai.grakn.test.GraphContext;
+
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Ignore;
